Stop using a leaked variable. (flake8)
scope is a variable that is defined lower in the for loop
and is leaked up in a future pass.

Also this code was slightly convoluted.
The goal is to stop rendering while the current scope is falsy.
Adding a new falsy scopes for every section inside
(So it doesn't start rendering on the next end tag).

I don't really know what I was thinking when I first coded this,
but all we have to do is add False instead of whatever in the
world I was doing before.