Add support for complex types
Before we were only allowing subscriptable structures,
now we can take something subscriptable or more complex
like a class with attributes.

This *shouldn't* cause problems with builtins, but I'm not 100% sure.
I've checked everything I could think of.

I might want to make this opt-in at somepoint, if some people are
having issues with it.